requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0
authlib==1.3.0
